Add in Shaped Charges. I've had this laying around so I thought I would spruce it up and release.

+++ src/main/java/com/grim3212/mc/pack/decor/block/BlockCraftBone.java
@@ -54,7 +54,7 @@ public class BlockCraftBone extends BlockManual {
 
 	@Override
 	public int getMetaFromState(IBlockState state) {
-		return (Integer) state.getValue(CYCLE);
+		return state.getValue(CYCLE);
 	}
 
 	@Override
